// src/auth/models/role.enum.ts
export enum Role {
  Read_all = 'read_all',
  Create_all = 'create_all',
  Update_all = 'update_all',
  Delete_all = 'delete_all',
  Read_restaurante = 'read_restaurante',
  Read_Producto = 'read_producto',
  Read_Cultura = 'read-cultura',
  Read_receta = 'read_receta',
  Read_pais = 'Read_pais',
}